% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bulltrout}
\alias{bulltrout}
\title{Bull Trout Counts}
\format{
A tbl data frame:
\describe{
  \item{Year}{The year (int).}
  \item{KasloRedds}{The Kaslo Creek redd count (int).}
  \item{KeenRedds}{The Keen Creek redd count (int).}
  \item{KasloCounter}{The Kaslo Creek resistivity counter count (int).}
}
}
\usage{
bulltrout
}
\description{
The annual resistivity counter and redd counts for Kaslo and Keen Creeks.
}
\seealso{
\code{\link{kootlake}}.
}
\keyword{datasets}
