% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishery_catch}
\alias{fishery_catch}
\title{Fishery Catch Counts}
\format{A data.frame
\describe{
 \item{StartYear}{The start year of the angling season (int).}
 \item{EndYear}{The end year of the angling season (int).}
 \item{Harvested}{Estimated count of individuals harvested (int).}
 \item{Released}{Estimated count of individuals released (int).}
 \item{MinWeight}{Bottom of weight range (int).}
 \item{MaxWeight}{Top of weight range (int).}
 \item{Species}{Species, RB or BT (factor. 2 levels).}
}}
\usage{
fishery_catch
}
\description{
Harvested and Released counts for Bull and Rainbow Trout
}
\seealso{
\code{\link{kootlake}}.
}
\keyword{datasets}
