% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gerrard}
\alias{gerrard}
\title{Gerrard Spawner Counts}
\format{A data.frame:
\describe{
  \item{Year}{The year of the spawner count as an integer.}
  \item{PeakCount}{The peak count as an integer.}
  \item{FishDays}{The total number of fish days as estimated by AUC assuming
  100 observer efficiency as an integer.}
}}
\usage{
gerrard
}
\description{
The Gerrard Rainbow Trout spawner counts from 1961 to 2014.
}
\details{
The total escapement can be estimated by multiplying the peak count by 3.08
or by dividing the total number of fish days by the residence time (currently)
assumed to be 11.9 days.
}
\seealso{
\code{\link{kootlake}}.
}
\keyword{datasets}

