% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_fishery}
\alias{get_fishery}
\title{Kootenay Lake Fishery}
\usage{
get_fishery()
}
\value{
The fishery tbl data frame.
}
\description{
Kootenay Lake Fishery
}
\examples{
get_fishery()
}
\seealso{
\code{\link{fishery}}
}
