% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_gerrard}
\alias{get_gerrard}
\title{Gerrard Escapement}
\usage{
get_gerrard()
}
\value{
The gerrard tbl data frame.
}
\description{
Gerrard Escapement
}
\examples{
get_gerrard()
}
\seealso{
\code{\link{gerrard}}
}
