% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_kokanee}
\alias{get_kokanee}
\title{Kokanee Escapement}
\usage{
get_kokanee()
}
\value{
The kokanee tbl data frame.
}
\description{
Kokanee Escapement
}
\examples{
get_kokanee()
}
\seealso{
\code{\link{kokanee}}
}
