% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-na.R
\name{kl_estimate_na}
\alias{kl_estimate_na}
\title{Estimate Missing Values}
\usage{
kl_estimate_na(x = kootlake::bulltrout, variable = "KasloRedds",
  predictor = "KasloCounter")
}
\arguments{
\item{x}{A data frame with columns of the variable and predictor.}

\item{variable}{A string of the name of the variable column.}

\item{predictor}{A string of the name of the predictor column.}
}
\value{
The original data frame with missing variable values replaced by the estimated value.
}
\description{
Estimates missing variable values from a predictor using a linear model.
}
\examples{
kl_estimate_na()
}
