% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kokanee}
\alias{kokanee}
\title{Kokanee Escapement}
\format{A data.frame:
\describe{
  \item{Year}{The year (int).}
  \item{Lardeau}{The Lardeau River escapement (int).}
  \item{MeadowCreek}{The Meadow Creek escapement  (int).}
  \item{Fecundity}{The mean number of eggs per female  (int).}
  \item{LengthMale}{The mean male spawner fork length in mm (int).}
  \item{LengthFemale}{The mean female spawner fork length in mm (int).}
}}
\usage{
kokanee
}
\description{
The Kokanee escapement.
}
\seealso{
\code{\link{kootlake}}.
}
\keyword{datasets}

