% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{temp_function}
\alias{temp_function}
\title{temp_function}
\usage{
temp_function(x)
}
\arguments{
\item{x}{Any object}
}
\value{
The same object.
}
\description{
function included as workaround for issue:
https://github.com/r-lib/covr/issues/427
}
\examples{

temp_function(3)

}
