% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fishery}
\alias{fishery}
\title{Kootenay Lake Fishery}
\format{A data.frame:
\describe{
  \item{Year}{The start of the angling season which runs from April to March. (int).}
  \item{AnglerHours}{The estimated number of angler hours expended by anglers who purchased a Kootenay Lake Rainbow Trout (KLRT) Conservation Surcharge License (int).}
}}
\usage{
fishery
}
\description{
The Kootenay Lake fishery data.
}
\seealso{
\code{\link{kootlake}}.
}
\keyword{datasets}

