% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{kootlake}
\alias{kootlake}
\alias{kootlake-package}
\title{Kootenay Lake Data}
\description{
Datasets for Kootenay Lake.
}
\details{
Currently just annual peak counts and fish days
for the number of Rainbow Trout spawners at Gerrard
as provided by the \code{\link{gerrard}} data.
}
\examples{
library(ggplot2)

data(gerrard)
gerrard$Escapement <- gerrard$PeakCount * 3.08

ggplot(data = gerrard, aes(x = Year, y = Escapement)) +
  geom_line() + expand_limits(y = 0)
}

